GUI: avoid calling Refresh() when nothing changes
Repainting the screen is expensive and causes lots of idle CPU usage when
the printer is continuously sending updates over the network.  In general,
we should avoid causing widgets to repaint if they don't change; we add some
checks to see if anything has actually changed inside of a widget before we
actually commit to repainting.